sna: Discard the mask for "non-overlapping" glyphs
If we can acheive the same rasterisation results without the mask,
rendering the glyphs-to-dst is so much faster that it outweighs the cost
of checking for overlapping glyphs.

The penalty is then for code that correctly declared that it required
a mask, who now have an extra ~10% overhead in the processing of their
glyphs.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>