sna/gen6: Bump the WM thread count to 80
Note that we should only do this when "WiZ Hashing" is disabled. So we
should be checking the GT_MODE register (bring on i915_read!) to be sure
that is safe to do so. However, it gives a big boost to performance of
render copies...  It also causes perf benchmarks to hit thermal limits
much quicker.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>