sna/traps: Translate the extents for the rasterization threads
The single-threaded code used the pre-computed width/height and only
required the origin from the bounds. However, the threads need to
allocate memory for themselves based on the computed bounds, and so it
helps if those bounds are then correct (rather than only the top-left
being in local space with the bottom-right in global coordinates).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>