sna/dri: Check rather than assert for a change in back/front sizes
The DRI2 midlayer may, or may not, recreate the back buffer after a
pixmap change. The result is that the back buffer may or may not be
flagged as unsuitable for flipping, so be paranoid and double check.
(This behaviour was first introduced in 1.9.0 and looks to be removed
again in 1.15.0)

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>