sna: Don't flag IGNORE_CPU for partial overwritten blocks
As we set the flags for the migration extents, but only operate upon the
boxes in the region, we need to be careful to be sure to migrate data
outside of the boxes overwritten.

Fixes regression from
commit 94cbe7bf7b7acc9a7f2cb56c5e275af028f3cdc8 [2.21.10]
Author: Chris Wilson <chris@chris-wilson.co.uk>
Date:   Thu Jun 20 19:40:44 2013 +0100

    sna: Mark overwriting CopyArea as not needing the dst damage

If we start passing around regions rather than extents, we coud do
finergrained migration decision.

Reported-by: Clemens Eisserer <linuxhippy@gmail.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=55244
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>