sna: Inspect the dirty boxes when querying whether damage contains a rectangle
This helps in the cases where we have subtracted a small number of
rectangles from an all-damage pixmap (such as a number of successive
GetImage, PutImage operations). The danger is that we end up searching a
long list of dirty boxes - maybe just search the first chunk if that
becomes noticeable?

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>