sna: Only flush after the BLT operation if we have more than 2 distinct bo
In order to preserve the optimisation of discarding incomplete batches,
we don't always want to immediately submit the batch after inserting the
first command. As we currently only cancel a batch if it only touches
the bo being discarded, we can skip the immediate flush if it only
accesses one bo and maybe be able to use the undo optimisation later.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>