sna/dri: Disable TripleBuffering by default for compositors
Ideally, the method of swapping is something that the applications have
control over, along with how to synchronise to the vertical refresh.
Whilst triple buffering is good to reduce jitter for games (at the cost of
an extra frame of latency, usually considered a good tradeoff), it
prevents the applications from accurately controlling the presentation
of animations. One vocal critique is Owen Taylor, who demands accurate
swap control for smooth animations in gnome-shell. For example,

http://blog.fishsoup.net/2012/11/28/avoiding-jitter-in-composited-frame-display/

In lieu of application control, just apply a quirk for the compositor.
Everyone else will just have to wait for DRI3.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>