sna/dri: Prefer the source ring
When copying from the DRI client onto the scanout, we return the source
buffer back to the client afterwards. Therefore if we force the source
to switch rings, we incur a double ring switch from the client to us and
back. If we force the dst to switch rings, it will likely be then
correct for all subsequent copies as well.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>