sna: Make sure we assert our Master status before copying the fbcon
Being able to read back the fbcon handle as a non-Master process is an
information leak that will be fixed. We should already be Master by
this point by virtue of the sequence in which we obtain the device fd.
However, to be pedagogically correct, call drmSetMaster() before the
fbcon copy.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>