sna: Allow user specification of number of VirtualHeads
Previously, we instantiated a fake output in case we had a machine with
no output. (For certain server-class products.) The Bumblee project were
also doing something very similar in order to fake an extended desktop
on the Intel igfx and copy portions onto a discrete GPU. (The preferred
method for doing this upstream is through the use of PRIME). As the code
is very similar, we can support both use-cases simultaneously.

This adds the option:

Section "Device"
  Driver "intel"
  Option "VirtualHeads" "<count>"
EndSection

to allow the user to specify an additional set of fake outputs, which
can then be controlled using xrandr.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>