intel-virtual-output: Decouple the fake mode before deleting
When we install a fake mode on the output, we have to perform a
round-trip for a reprobe. During that time, some other client (such as
gnome-shell) will be notify of the change in output status and may
attempt to configure it. This leaves us unable to delete the mode as it
is currently active - so first disable the new head and then delete the
mode.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>