sna: Include Pineview in fence over-allocation checks
When creating the object, we need to take account that fenced regions
in the aperture are in power-of-two sizes and so can be much larger than
the actual object. This includes Pineview, which although is more
relaxed with regards to fence alignment of GPU operations, still imposes
the same restrictions on the fence registers.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>